+++ CHANGES.md
@@ -5,6 +5,7 @@
 ##### Fixes :wrench:
 
 - Fixed an issue where tileset styles would be reapplied every frame when a tileset has a style and `tileset.preloadWhenHidden` is true and `tileset.show` is false. Also fixed a related issue where styles would be reapplied if the style being set is the same as the active style. [#9223](https://github.com/CesiumGS/cesium/pull/9223)
+- Fixed an issue where tileset stops refining when empty tile is near the leaf of the tileset. [#9140](https://github.com/CesiumGS/cesium/pull/9140)
 
 ### 1.75 - 2020-11-02